.ipynb_checkpoints/requirements-checkpoint.txt
streamlit>=1.32
openai>=1.0.0
pandas
python-docx
pdfplumber
openpyxl

